refactor using ActionMenu & Delete scenario
